created User class and tested with Dataset
